Create sql-database-query
How we created a table that summarised the user activity. We were not using a sign up/log in system so we wanted to understand more about the unique users. This is because we send leads of buyers and sellers out to our clients and we needed to ensure that we had legitimate details and we didn't send out leads twice.